Fixed it crashed while saving or opening an attachment in an IMAP folder with FLAG_CACHEWHENREAD flag.
While it opens or saves an attachment, it calls MessageContext::getMessage which will get a message from the server. While getting a message, it updates the current message if the folder has FLAG_CACHEWHENREAD flag and the whole message has not been cached, which will update keys of MessageHolder. When keys of MessageHolder are updated, MessageModel updates the current message and deletes the old current MessageContext. To put it all together, calling MessageContext::getMessage will delete itself which caused the problem. To avoid this problem, we have to copy the current MessageContext before calling getMessage in certain situations.

In addition to this problem, I changed Account to update MessageHolder's flags in Account::updateMessage. This is because the flags should be updated when its keys have been changed, so that observers of changes on keys can see the latest flags.


git-svn-id: https://subversion.assembla.com/svn/q3/trunk@4396 8af8166b-12a6-a448-a533-9086ace3f9f6